improve: Relax the signature check in BatchVote (#170)
## What ❔

`BatchVote::update` ignores invalid signatures instead of returning an
error.

## Why ❔

Because on `stage-v2` the `moby` variant of the external node runs an
earlier version of the code without the `Batch::genesis` field, which
causes it to reject all votes coming from the main and the other
external nodes. The error causes it to disconnect its gossip peers,
which is a bit heavy handed given that this is an unstable feature.