Pin Rust version to 1.79 (#177)
## What ❔

Created a `rust-toolchain` file pinning it to the stable version when
1.79 was released.
Reverted the change to `stream.rs` brought about by 1.80 in
https://github.com/matter-labs/era-consensus/pull/176
Bumps the version to `0.1.0-rc.8`

## Why ❔

Because the `zksync-era` build [doesn't
work](https://github.com/matter-labs/zksync-era/actions/runs/10249428261/job/28352782628?pr=2583)
with `size_of` without qualification like `core::mem::` or `std::mem::`.

`zksync-era` is pinned at `nightly-2024-05-07` at the moment. I tried to
update it to the last nightly where it was still Rust 1.80, which was
`nightly-2024-06-08` but it cannot compile `boojum`, nor does it compile
with a later nightly I tried where it's already Rust 1.82.