feat: Implement ChonkyBFT (#211)
Implemented types and logic for ChonkyBFT. Note that I'm not keeping the
old types, that's because `ReplicaCommit` has no changes to it (I
checked and serialization is the same for old and new, old signatures
still work) as well as `CommitQC` and `FinalBlock`. Any type that an EN
could receive, or that ends on our database, doesn't suffer any
alterations. So it seems that the upgrade to ChonkyBFT is
backwards-compatible for ENs. Since we only have one validator, we can
upgrade the whole network without a planned hard fork.
Fixes BFT-452