feat: Add versioning and protocol version 2 (in development) (#237)
## What ❔

This adds a new protocol version and the ability to switch protocol
versions on regenesis (i.e. when the genesis spec is changed on the main
node consensus config). The refactor makes it simpler to add/deprecate
versions in the future.
The new protocol version just adds a new field, `epoch_number`, to the
`View` struct. It isn't currently used, but it will be used in future
PRs to introduce validator committee rotation. For now, it will just be
used to test the protocol upgrade.

## Why ❔

We need to be able to upgrade the consensus protocol.