refactor: Remove attester code (#248)
## What ❔

It deprecates all the code related to the attester role. Since ZK OS
won't have L1 batches, this role won't be needed in the future. Plus, if
we decide to do L1 consensus verification before ZK OS, it's simpler to
just do an operator multisig (the functionality is reduced but it's a
temporary solution anyway).