feat: Improve leader selection (#251)
## What ❔

Allows us to only rotate the leader every N blocks, instead of every
block. This will allow us (when we have leader rotation) to pipeline the
block production with the block validation + consensus and so only one
VM execution would be sequential.
If the current leader fails a single block, we'll view change to a
different leader instead of letting all N slots timeout. With a very
large N, this is effectively equal to having a fixed leader (that
changes on failure).

## Why ❔

Doing this now because we want the ConsensusRegistry contract to already
pick a leader instead of letting it exist in the genesis like now
(because having the nodes getting the validator set from the contract
but the leader from the genesis file is a huge footgun). Given this, we
might as well already implement the ideal leader selection we want now
instead of doing a temporary fix.

Part of PLA-1122